Add Code Collapse
Hi,

I created this simple, straightforward extension to improve quality of life for all GitHub users :) How do you feel about adding it?

I plan on publishing a Chrome version, so if you feel like validating the PR but would rather wait for the Chrome version to be available, that's fine. And no hard feelings if you're not interested with the extension itself of course ;)

Thank you for the repository btw.